+++ CHANGELOG.md
@@ -35,6 +35,8 @@ guarantee message expiration
 - `contracts.process.transaction` function processing transaction to check errors and get output
 - `contracts.run.local` and `contracts.run.local` functions now have `fullRun` flag to emulate
 node transaction processing and calculate fees
+- `tonsdk` command line tool.
+- `ton_client` function `get_method_names`.
 
 ## May 22, 2020
 #### Fix